+++ example/example.go
@@ -2,7 +2,7 @@ package example
 
 import (
 	"fmt"
-	"go-xml-reporter/junit"
+	observabilityReporter "go-xml-reporter/observability-reporter"
 	"log"
 	"os"
 
@@ -21,23 +21,23 @@ func Example() {
 	BSTACK_USERNAME := os.Getenv("BSTACK_USERNAME")
 	BSTACK_PASSWORD := os.Getenv("BSTACK_PASSWORD")
 
-	// Create a new junit reporter object
-	reporter1 := junit.NewJUnitReporter()
+	// Create a new reporter object
+	build1 := observabilityReporter.NewJUnitReporter()
 
 	// Adding build info
-	buildIdentifier, err := reporter1.CreateBuild(BSTACK_USERNAME, BSTACK_PASSWORD, "test-build", "Junit report uploads - Go XML Library", "", []string{"junit_upload", "regression"}, "http://localhost:8080/", "mocha, 10.0.0", map[string]string{}, map[string]string{})
+	buildIdentifier, err := build1.CreateBuild(BSTACK_USERNAME, BSTACK_PASSWORD, "test-build", "Junit report uploads - Go XML Library", "", []string{"junit_upload", "regression"}, "http://localhost:8080/", "mocha, 10.0.0", map[string]string{}, map[string]string{})
 	if err != nil {
 		panic(err)
 	}
 
 	// Updating build info
-	_, updateBuildErr := reporter1.UpdateBuildDetails(buildIdentifier, []string{"xml"}, "", "", map[string]string{}, map[string]string{})
+	_, updateBuildErr := build1.UpdateBuildDetails(buildIdentifier, []string{"xml"}, "", "", map[string]string{}, map[string]string{})
 	if updateBuildErr != nil {
 		panic(updateBuildErr)
 	}
 
 	// Adding a test
-	testIdentifier1, addTestErr1 := reporter1.AddTestRun(buildIdentifier, "Test 1", "failed",
+	testIdentifier1, addTestErr1 := build1.AddTestRun(buildIdentifier, "Test 1", "failed",
 		`Test 1 error stacktrace 1`, "2023-05-24T11:00:14", "3.133", "nil",
 		`classname 1`,
 		`/Users/testuser/work/samples/signIn.e2e.js`,
@@ -56,7 +56,7 @@ func Example() {
 	}
 
 	// Updating a test
-	_, updateTestErr := reporter1.UpdateTestRunDetails(testIdentifier1, map[string]interface{}{
+	_, updateTestErr := build1.UpdateTestRunDetails(testIdentifier1, map[string]interface{}{
 		"tag":        []string{"extra"},
 		"attachment": []string{"example/screenshots/checklist.png"},
 	}, "log info updated", map[string]string{})
@@ -65,7 +65,7 @@ func Example() {
 		panic(updateTestErr)
 	}
 
-	firstSplitResponse, firstSplitErr := reporter1.SendReport(buildIdentifier)
+	firstSplitResponse, firstSplitErr := build1.SendReport(buildIdentifier)
 	if firstSplitErr != nil {
 		panic(firstSplitErr)
 	}
@@ -73,12 +73,12 @@ func Example() {
 	fmt.Println("First split response message: ", firstSplitResponse)
 
 	// Adding more tests
-	reporter1.AddTestRun(buildIdentifier, "Test 1", "passed",
+	build1.AddTestRun(buildIdentifier, "Test 1", "passed",
 		``, "2023-05-24T11:00:14", "3.133", "nil",
 		`classname 1`,
 		`/Users/testuser/work/samples/home.js`,
 		nil, "log info", nil)
-	reporter1.AddTestRun(buildIdentifier, "Test 2",
+	build1.AddTestRun(buildIdentifier, "Test 2",
 		"failed", `first error stacktrace`,
 		"2023-05-24T11:00:17",
 		"2.343",
@@ -87,7 +87,7 @@ func Example() {
 		`/Users/testuser/work/samples/home.js`,
 		nil, "", nil,
 	)
-	reporter1.AddTestRun(buildIdentifier, `BStack - Login fucntionality &quot;after all&quot; hook for &quot;Login with invalid credentials - Dynamic Skip&quot;`,
+	build1.AddTestRun(buildIdentifier, `BStack - Login fucntionality &quot;after all&quot; hook for &quot;Login with invalid credentials - Dynamic Skip&quot;`,
 		"failed", `second error stacktrace`,
 		"2023-05-24T11:00:17",
 		"1.500",
@@ -97,19 +97,19 @@ func Example() {
 		nil, "", nil,
 	)
 
-	secondSplitResponse, secondSplitErr := reporter1.SendReport(buildIdentifier)
+	secondSplitResponse, secondSplitErr := build1.SendReport(buildIdentifier)
 	if secondSplitErr != nil {
 		panic(secondSplitErr)
 	}
 
 	fmt.Println("Second split response message: ", secondSplitResponse)
 
-	reporter1.AddTestRun(buildIdentifier, "Test 1", "failed",
+	build1.AddTestRun(buildIdentifier, "Test 1", "failed",
 		`Test 1 error stacktrace 1`, "2023-05-24T11:00:14", "3.133", "parentsuite 1",
 		`parentsuite 1 classname 1`,
 		`/Users/testuser/work/samples/signIn.e2e.js`,
 		nil, "", nil)
-	reporter1.AddTestRun(buildIdentifier, "Test 2",
+	build1.AddTestRun(buildIdentifier, "Test 2",
 		"failed", `first error stacktrace`,
 		"2023-05-24T11:00:17",
 		"2.343",
@@ -118,7 +118,7 @@ func Example() {
 		`/Users/testuser/work/samples/signIn.e2e.js`,
 		nil, "", nil,
 	)
-	reporter1.AddTestRun(buildIdentifier, `BStack - Login fucntionality &quot;after all&quot; hook for &quot;Login with invalid credentials - Dynamic Skip&quot;`,
+	build1.AddTestRun(buildIdentifier, `BStack - Login fucntionality &quot;after all&quot; hook for &quot;Login with invalid credentials - Dynamic Skip&quot;`,
 		"failed", `second error stacktrace`,
 		"2023-05-24T11:00:17",
 		"1.500",
@@ -129,7 +129,7 @@ func Example() {
 	)
 
 	// Stop a build ( this will generate xml report for tests added and upload it to browserstack observability. )
-	respMessage, err := reporter1.StopBuild()
+	respMessage, err := build1.StopBuild()
 	if err != nil {
 		panic(err)
 	}
