+++ observability-reporter/stopBuild.go
@@ -1,8 +1,6 @@
-package junit
+package observabilityReporter
 
-import "errors"
-
-func (jr *JUnitReporter) SendReport(buildIdentifier string) (string, error) {
+func (jr *JUnitReporter) StopBuild() (string, error) {
 
 	if len(jr.testSuites) > 0 {
 		err := generateXMLFromTestSuites(jr.buildDetails.buildIdentifier, jr.testSuites)
@@ -18,16 +16,13 @@ func (jr *JUnitReporter) SendReport(buildIdentifier string) (string, error) {
 		}
 
 		// Delete the created file & reset the inmemory to empty(default) values
-		jr.resetTestSuites()
-
-		// Create a folder to store xml files and attachments
-		createBuildDirErr := createBuildDirectory(jr.buildDetails.buildIdentifier)
-		if createBuildDirErr != nil {
-			removeBuildAssets(jr.buildDetails.buildIdentifier)
-		}
+		jr.resetData()
 
 		return respMessage, nil
 	}
 
-	return "", errors.New("no tests are added to builder to process xml report")
+	// Delete the created file & reset the inmemory to empty(default) values
+	jr.resetData()
+
+	return "Build stop is successful", nil
 }
